vimrc: better indicate inactive panes in tmux
we often/usually run vim in a tmux, and we want to more easy tell which
pane is currently active. by setting tmux background color differently
for active/inactive panes, and making the vim background color just the
background color of tmux, this becomes possible!

based on this SO answer:
https://stackoverflow.com/questions/37712730/set-vim-background-transparent